+++ packages/web-components/src/components/global-header/components/global-header/src/types/Header.types.ts
@@ -390,3 +390,9 @@ declare global {
     _solis: solisWindowConfig;
   }
 }
+
+export interface solisSessionManagerConfig {
+  tokenRefreshInterval?: number;
+  idleTimeoutInterval?: number;
+  basePath?: string;
+}